+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -10,3 +10,5 @@ pub mod identifiers;
 pub mod import_directives;
 pub mod literals;
 pub mod modifiers;
+pub mod pragma_directives;
+pub mod source_units;
