+++ aderyn_core/src/ast/impls/node/pragma_directives.rs
@@ -1,15 +1,6 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
-
-#[derive(Default, Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct PragmaDirective {
-    pub literals: Vec<String>,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Node for PragmaDirective {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
