+++ aderyn_core/src/ast/mod.rs
@@ -3,8 +3,6 @@ pub mod impls;
 pub mod macros;
 pub mod magic;
 pub mod node_type;
-pub mod pragma_directives;
-pub mod source_units;
 pub mod statements;
 pub mod structures;
 pub mod types;
@@ -14,9 +12,8 @@ pub mod variables;
 pub mod yul;
 
 pub use self::{
-    impls::*, magic::*, node_type::*, pragma_directives::*, source_units::*, statements::*,
-    structures::*, types::*, user_defined_value_types::*, using_for_directives::*, variables::*,
-    yul::*,
+    impls::*, magic::*, node_type::*, statements::*, structures::*, types::*,
+    user_defined_value_types::*, using_for_directives::*, variables::*, yul::*,
 };
 
 pub use ast_nodes::*;
